Subscriptions widget added to GoalEditTimeframePage
